Implemented segmentation
The OS now sets up its own segmentation

These additions break the OS currently, and I have not figured out why yet. I have
successfully tested the segmentation code and know that it works, so I wanted to
commit it and allow others to pull these changes and figure out why the OS won't
start anymore. I believe it's an executable size problem, because as soon as I strip
out the serial code, it works. That's how I was able to test the segmentation code.
And, the error Bochs gives is that the executable is invalid.

Signed-off-by: Alexander Anderson <atanderson61@outlook.com>